Run CI only once on merge commits.
Currently, we run the CI as MR pipelines, i.e. the CI is run on a temporary merge result of the source and target branch. Additionally, the CI is run on the actual source/target merge result, such that the same tests are done twice.

This changes this, by not running the CI on the protected branch (master).