Created a Task baseclass and registered run_command with @celery_app.task.
* This required a reversal of the import (registration not anymore in celery_worker.py, where it anyway would not make sense, as soon as multiple task types come up (soon)).
* run_command tests needed file-based config with WESKIT_CONFIG being set.

Additionally, I changed the way how the tests/remote.yaml works. The current way is a PITA and results in multiple check-ins of passwords.

Now, there is a `remote-template.yaml`. The actual `remote.yaml` that is used in the tests, is included in the `.gitignore`.

When setting up the config for testing, copy the `remote-template.yaml` to `remote.yaml`, edit it and done. No chance of erroneously checking in passwords.